Pig Exercise
The spreadsheet Pig_Labs_Whitehouse.xlsx has the PigLatin queries to ingest raw Whitehouse visits data into HDFS and then select the columns that are needed for further analysis. 
The Whitehouse data was analysed to see among all the visitors the group of visitors that visited POTUS and their count.